pyproject.toml
[tool.poetry]
name = "ll-datasets"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "ll_datasets"}]

[tool.poetry.dependencies]
python = "^3.12"
mkdocs = "^1.5.3"
mkdocstrings = "^0.24.3"
mkdocstrings-python = "^1.9.2"
mkdocs-material = "^9.5.17"
transformers = "^4.41.0"
gdown = "^5.2.0"
python-dotenv = "^1.0.1"
pandas = "^2.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
